add BaseDAO, CouponDAO, UserDAOImpl. Fix entity and dao interfaces

+++ FinalTask/src/main/java/by/training/final_task/dao/interfases/CategoryDAO.java
@@ -6,8 +6,8 @@ import by.training.final_task.exception.PersistentException;
 import java.util.List;
 
 public interface CategoryDAO {
-    Category getById(int id) throws PersistentException;
+    Category getById(long id) throws PersistentException;
     List<Category> getAll() throws PersistentException;
-    boolean delete(int id) throws PersistentException;
+    boolean delete(long id) throws PersistentException;
     int GetAmountOfCategory() throws PersistentException;
 }
