+++ FinalTask/src/main/java/by/training/final_task/dao/interfases/CompanyProviderDAO.java
@@ -6,9 +6,9 @@ import by.training.final_task.exception.PersistentException;
 import java.util.List;
 
 public interface CompanyProviderDAO extends DAO<CompanyProvider> {
-    CompanyProvider getById(int id) throws PersistentException;
+    CompanyProvider getById(long id) throws PersistentException;
     CompanyProvider getByAddress(String address) throws PersistentException;
     List<CompanyProvider> getAll() throws PersistentException;
-    boolean delete(int id) throws PersistentException;
+    boolean delete(long id) throws PersistentException;
     int getAmountOfCompany() throws PersistentException;
 }
