+++ FinalTask/src/main/java/by/training/final_task/dao/interfases/CouponUserDAO.java
@@ -7,18 +7,18 @@ import java.util.Date;
 import java.util.List;
 
 public interface CouponUserDAO extends DAO<CouponUser> {
-    CouponUser getById(int id) throws PersistentException;
+    CouponUser getById(long id) throws PersistentException;
     //???
     List<CouponUser> getAll(int offset, int limit) throws PersistentException;
-    List<CouponUser> getAllCouponsCurrentUser(int userId, int offset, int limit)
+    List<CouponUser> getAllCouponsCurrentUser(long userId, int offset, int limit)
         throws PersistentException;
-    List<CouponUser> getAllBetweenDatesCurrentUser(int userId, Date startDate,
+    List<CouponUser> getAllBetweenDatesCurrentUser(long userId, Date startDate,
                                           Date endDate, int offset, int limit)
             throws PersistentException;
-    int getCountCouponNameCurrentUser(int userId, int offset, int limit)
+    int getCountCouponNameCurrentUser(long userId, int offset, int limit)
             throws PersistentException;
-    int getCountBetweenDatesCurrentUser(int userId, Date startDate,
+    int getCountBetweenDatesCurrentUser(long userId, Date startDate,
                                     Date endDate) throws PersistentException;
     int getAllCount() throws PersistentException;
-    boolean delete(int id) throws PersistentException;
+    boolean delete(long id) throws PersistentException;
 }
