+++ FinalTask/src/main/java/by/training/final_task/dao/interfases/ReviewsDAO.java
@@ -6,12 +6,12 @@ import by.training.final_task.exception.PersistentException;
 import java.util.List;
 
 public interface ReviewsDAO extends DAO<Reviews> {
-    Reviews getById(int id) throws PersistentException;
+    Reviews getById(long id) throws PersistentException;
     List<Reviews> getAll(int offset, int limit) throws PersistentException;
-    List<Reviews> getAllReviewsCurrentUser(int userId, int offset, int limit)
+    List<Reviews> getAllReviewsCurrentUser(long userId, int offset, int limit)
         throws PersistentException;
     int getAllCount() throws PersistentException;
-    int getCountReviewsCurrentUser(int userId, int offset, int limit)
+    int getCountReviewsCurrentUser(long userId, int offset, int limit)
         throws PersistentException;
-    boolean delete(int id) throws PersistentException;
+    boolean delete(long id) throws PersistentException;
 }
